Move files to backend folder
