+++ src/assets/js/backend/functions.js
@@ -1,0 +1,11 @@
+const functions = {
+  rgbToHex(rgb) {
+    let hex = Number(rgb).toString(16);
+    if (hex.length < 2) {
+      hex = `0${hex}`;
+    }
+    return hex;
+  },
+};
+
+export default { functions };
